requirements.txt
rembg>=2.0.0
Pillow>=9.0.0
numpy>=1.21.0
onnxruntime>=1.12.0